requirements.txt
matplotlib==3.7.2
numpy==1.24.4
onnx==1.17.0
onnxruntime==1.15.1
pandas==2.0.3
Pillow==10.0.0
torch==2.7.0+cu118
torchvision==0.22.0+cu118
umap_learn==0.5.9.post2
